requirements.txt
pydantic==1.10.14
email-validator==2.1.0.post1
tortoise-orm==0.18.1
aerich == 0.6.2
fastapi==0.109.2
passlib==1.7.4
python-jose[cryptography]
bcrypt==4.0.1
pytest==7.2.0
httpx==0.23.3
pytest-asyncio==0.20.3
uvicorn==0.15.0
python-multipart==0.0.5
asyncpg==0.27.0
